Resolved python/netsnmp/client_intf.c by copying from 2010.snmp's
previous 5.7.2rc2 integration (the "old way").  Re-resolved everything
else, largely in favor of upstream.

Conflicts:
	agent/Makefile.in
	agent/mibgroup/agentx/subagent.c
	agent/mibgroup/host/data_access/swrun_procfs_status.c
	agent/mibgroup/ucd-snmp/pass_persist.c
	agent/snmp_agent.c
	apps/Makefile.in
	configure
	configure.d/config_os_functions
	include/net-snmp/net-snmp-config.h.in
	local/pass_persisttest
	python/netsnmp/client_intf.c
	snmplib/int64.c
	snmplib/snmp_alarm.c
	snmplib/snmp_client.c
	snmplib/snmpv3.c
	snmplib/tools.c
	snmplib/transports/snmpUDPDomain.c

+++ agent/mibgroup/disman/schedule/schedCore.c
@@ -302,15 +302,14 @@ sched_nextTime( struct schedTable_entry *entry )
              entry->schedNextRun = now + entry->schedInterval;
         }
         DEBUGMSGTL(("disman:schedule:time", "periodic: (%ld) %s",
-                                  entry->schedNextRun,
-                           ctime(&entry->schedNextRun)));
+                    (long) entry->schedNextRun, ctime(&entry->schedNextRun)));
         break;
 
     case SCHED_TYPE_ONESHOT:
         if ( entry->schedLastRun ) {
             DEBUGMSGTL(("disman:schedule:time", "one-shot: expired (%ld) %s",
-                                  entry->schedNextRun,
-                           ctime(&entry->schedNextRun)));
+                        (long) entry->schedNextRun,
+                        ctime(&entry->schedNextRun)));
             return;
         }
         /* Fallthrough */
@@ -425,8 +424,7 @@ sched_nextTime( struct schedTable_entry *entry )
          */
         entry->schedNextRun = mktime( &next_tm );
         DEBUGMSGTL(("disman:schedule:time", "calendar: (%ld) %s",
-                                  entry->schedNextRun,
-                           ctime(&entry->schedNextRun)));
+                    (long) entry->schedNextRun, ctime(&entry->schedNextRun)));
         return;
 
     default:
